частично заменил на классы
