+++ src/utils/helpers.js
@@ -26,7 +26,6 @@ export const createElement = (template) => {
   return newElement.firstChild;
 };
 
-
 export const getRandomInteger = (a = 0, b = 1) => {
   const lower = Math.ceil(Math.min(a, b));
   const upper = Math.floor(Math.max(a, b));
