--- src/view/films-list-view.js
@@ -1,8 +1,0 @@
-export const createFilmTemplate = () => (
-  `<section class="films">
-    <section class="films-list">
-      <h2 class="films-list__title visually-hidden">All movies. Upcoming</h2>
-      <div class="films-list__container"></div>
-    </section>
-  </section>`
-);
