+++ src/view/films-view.js
@@ -1,0 +1,30 @@
+import { createElement } from '../utils/helpers.js';
+
+const createFilmTemplate = () => (
+  `<section class="films">
+    <section class="films-list">
+      <h2 class="films-list__title visually-hidden">All movies. Upcoming</h2>
+      <div class="films-list__container"></div>
+    </section>
+  </section>`
+);
+
+export default class FilmContainerView {
+  #element = null;
+
+  get element() {
+    if (!this.#element) {
+      this.#element = createElement(this.template);
+    }
+
+    return this.#element;
+  }
+
+  get template() {
+    return createFilmTemplate();
+  }
+
+  removeElement() {
+    this.#element = null;
+  }
+}
