+++ src/view/loading-view.js
@@ -1,10 +1,14 @@
 import { createElement } from '../utils/helpers.js';
 
-const createShowMoreTemplate = () =>
-  '<button class="films-list__show-more">Show more</button>';
-
-export default class сreateShowMore {
+const createLoadingTemplate = () => (
+  '<h2 class="films-list__title">Loading...</h2>'
+);
+export default class LoadingView {
   #element = null;
+  #cards = null;
+  constructor(cards) {
+    this.#cards = cards;
+  }
 
   get element() {
     if (!this.#element) {
@@ -15,11 +19,10 @@ export default class сreateShowMore {
   }
 
   get template() {
-    return createShowMoreTemplate();
+    return createLoadingTemplate(this.#cards);
   }
 
   removeElement() {
-    this.#element.remove();
     this.#element = null;
   }
 }
