+++ src/view/profile-view.js
@@ -1,4 +1,5 @@
 import { FAN, NOVICE } from '../utils/consts.js';
+import { createElement } from '../utils/helpers.js';
 
 const createRatingTemplate = (count) => {
   const getRank = () => {
@@ -14,11 +15,38 @@ const createRatingTemplate = (count) => {
   return `<p class="profile__rating">${getRank(count)}</p>`;
 };
 
-export const createProfileTemplate = (count) => {
+const createProfileTemplate = (count) => {
   const rating = count > 0 ? createRatingTemplate(count) : '';
 
-  return`<section class="header__profile profile">
-  ${rating}
-    <img class="profile__avatar" src="images/bitmap@2x.png" alt="Avatar" width="35" height="35">
-  </section>`;
+  return (
+    `<section class="header__profile profile">
+      ${rating}
+      <img class="profile__avatar" src="images/bitmap@2x.png" alt="Avatar" width="35" height="35">
+    </section>`
+  );
 };
+
+export default class ProfileView {
+  #element = null;
+  #count = null;
+
+  constructor(count) {
+    this.#count = count;
+  }
+
+  get template() {
+    return createProfileTemplate(this.#count);
+  }
+
+  get element() {
+    if (!this.#element) {
+      this.#element = createElement(this.template);
+    }
+
+    return this.#element;
+  }
+
+  removeElement() {
+    this.#element = null;
+  }
+}
