+++ src/view/sort-view.js
@@ -1,7 +1,34 @@
-export const createSortTemplate = () => (
+import { createElement } from '../utils/helpers.js';
+
+const createSortViewTemplate = () => (
   `<ul class="sort">
     <li><a href="#" class="sort__button sort__button--active">Sort by default</a></li>
     <li><a href="#" class="sort__button">Sort by date</a></li>
     <li><a href="#" class="sort__button">Sort by rating</a></li>
   </ul>`
 );
+
+export default class SortView {
+  #element = null;
+  #filters = null;
+
+  constructor(filters) {
+    this.#filters = filters;
+  }
+
+  get element() {
+    if (!this.#element) {
+      this.#element = createElement(this.template);
+    }
+
+    return this.#element;
+  }
+
+  get template() {
+    return createSortViewTemplate(this.#filters);
+  }
+
+  removeElement() {
+    this.#element = null;
+  }
+}
